Fixed Product Card and List Stories
